requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
